requirements.txt
boto3==1.34.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
